Styles: Create a new service for managing styles
This service will be able to check if the user has changed the app theme both from settings or from GNOME Tweaks, and then load the right style. And if the theme setted is a custom theme, the service will remove all the styles, since we don't know if the custom theme actually supports our styles.