symbolic icons: moved to its right location
Since GTK 4, now the symbolic icons need to respect fd.o path in order to be
used as part of the icon theme.

I was using the alias attribute on the gresources xml as a workaround to this
but it requires to write the full path of the icon twice, which is tiring to do
and makes the xml harder to read.

This commits moves all the icons to its right location inside of the 'data' dir
and adds a new gresources xml file to add them as an input for the compiler.