thumbnail: make session a private instance field
Since libsoup does not process each asynchronous request in parallel,
but queues them, it is vital that the thumbnails do not share the same
SoupSession instance.

This commit moves `_session` to be an instance field instead of a static
one, so that each thumbnail can be downloaded in parallel to the others,
speeding up the loading of the thumbnails.